+++ src/cli/send/plugin/bailOnFailedTestInterceptor.ts
@@ -6,7 +6,9 @@ export const bailOnFailedTestInterceptor = {
   id: 'bailOnFailed',
   afterTrigger: async function bail(hookContext: HookTriggerContext<[models.ProcessorContext], boolean>) {
     const context = hookContext.args[0];
-    const failedTest = context.httpRegion.testResults?.find?.(obj => !obj.result);
+    const failedTest = context.httpRegion.testResults?.find?.(obj =>
+      [models.TestResultStatus.FAILED, models.TestResultStatus.ERROR].includes(obj.status)
+    );
     if (failedTest) {
       throw failedTest.error || new Error('bail on failed test');
     }