+++ src/models/logHandler.ts
@@ -24,7 +24,6 @@ export interface LogHandler {
 }
 
 export interface ConsoleLogHandler extends LogHandler {
-  logTest(result: boolean, message: string): void;
   logPriority(message: string): void;
   collectMessages(): void;
   flush(): void;