+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - variables in gql Query body are replaced (AnWeber/vscode-httpyac#303)
 - allow defaultHeaders to overwrite accept or user-agent header
 - only try loading grpc reflection if no other protodefinition file is imported or it is explicitly triggerd using grpcReflection (#757)
+- add better handling for disabled and error tests (Anweber/vscode-httpyac#297, #760)
 
 ## [6.14.0] ( 2024-06-01)
 ### Features